nbs_cli/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "nbs-cli"
version = "1.0.0"
description = "Nitrux Bootstrap System — Lightweight command-line tool for building ISO files."
authors = [
    { name = "Uri Herrera", email = "uri_herrera@nxos.org" }
]
license = { text = "BSD-3-Clause" }
readme = "README.md"
requires-python = ">=3.6"
dependencies = [
    "typer[all]",
    "rich",
    "pyyaml",
    "requests",
    "python-debian",
    "tqdm"
]

[project.urls]
Homepage = "https://github.com/Nitrux/nbs"

[project.scripts]
nbs-cli = "nbs_cli.cli:main"
